Add SQL query to retrieve project-related user tasks
This commit adds a new SQL query to retrieve the tasks associated with a
user and a project. The query is implemented in the manager and is used by
the updated controller to return the relevant tasks. Additionally, a new
route has been added to route.js to support this functionality.